test(wgpu): split shadow_pass tests into sibling file
CI flagged shadow_pass.rs at 550 lines, over the 500-line cap from
scripts/check-rs-line-limit.sh. Move the 6 pure-logic regression tests
into shadow_pass_tests.rs and gate it with #[cfg(test)] in mod.rs.

shadow_pass.rs is back to 447 lines; tests still pass with the same
names so CI history stays continuous.
